- Added explicit cast in CoreAudio driver to fix debug build with modern XCode 8

+++ mt32emu_qt/src/audiodrv/CoreAudioDriver.cpp
@@ -81,7 +81,7 @@ bool CoreAudioStream::start() {
 
 	qDebug() << "CoreAudio: using default audio output device";
 
-	AudioStreamBasicDescription dataFormat = {sampleRate, kAudioFormatLinearPCM, kAudioFormatFlagsNativeEndian | kAudioFormatFlagIsSignedInteger, 4, 1, 4, 2, 16, 0};
+	AudioStreamBasicDescription dataFormat = {(Float64)sampleRate, kAudioFormatLinearPCM, kAudioFormatFlagsNativeEndian | kAudioFormatFlagIsSignedInteger, 4, 1, 4, 2, 16, 0};
 	OSStatus res = AudioQueueNewOutput(&dataFormat, renderOutputBuffer, this, NULL, NULL, 0, &audioQueue);
 	if (res || audioQueue == NULL) {
 		qDebug() << "CoreAudio: AudioQueueNewOutput() failed with error code:" << res;
